feat: Add role field to createResourceRole API endpoint
This commit adds the `role` field to the `createResourceRole` API endpoint in the `resourceRouter`. The `role` field is now included in the input object and is used to specify the role associated with the created resource role. This change enhances the functionality of creating resource roles and improves the accuracy of role assignment.